[refactor] 레포지토리 분리
본래 유저 레포지토리의 메서드 시그니처만 인터페이스로 분리하고 구현 내용을 레포지토리 구현체 클래스를 따로 만들었습니다.
인터페이스에는 @Repository 어노테이션을 부착하고 구현체에는 @RepositoryImpl 어노테이션을 부착하였습니다.
DependencyFactory 가 해당 어노테이션을 파악하여 의존성을 관리해줄 것을 기대합니다.
